Working on the third major attempt at this. Orchestration seemed to work well, but we are running too many duplicate assessments of the same information and it was exploding token usage. Now we are going to try something else.
I am going to use Tool Master + Igor + Lore Master and eliminate the Orchestration Step. WE are going to lean on Tool Master a bit more. We are going to allow Tool Master to assign tokens to the tools. Each token represents a query, so more tokens = more information queried.
We are then going to use one or more Igor to summarize the information, and then pass everything to Lore Master to assess.

To this end, I am designing a new Tool Master Token prompt, and a new TINY orchestration run